added autoconfig for ImageResize addon
Themes have a lexy extension with a configuration, that is similar to the ImageResize profiles. The autoconfig fires only if no ImageResize profiles were set manually.

So if nothing is configured and both MPGUI and ImgResize addon are installed, the predefined theme img sizes are used as img profiles to create thumbnails etc. on the fly after uploading new assets.